src/governance-worker/pyproject.toml
[project]
name = "apic-vibe-portal-governance-worker"
version = "0.0.0"
description = "Azure Container Job — Governance snapshot pipeline for APIC Vibe Portal"
requires-python = ">=3.14"
dependencies = [
    "apic-client",
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.21.0",
    "azure-monitor-opentelemetry>=1.6.4",
    "opentelemetry-api>=1.27.0",
    "pydantic-settings>=2.13.1",
    "structlog>=25.5.0",
]

[tool.uv.sources]
apic-client = { path = "../apic_client", editable = true }

[dependency-groups]
dev = [
    "ruff>=0.7.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["governance_worker"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::SyntaxWarning",
]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
